Add files via upload
For OffDumpGoal wrote code to get the bot off the ramp, orient itself, go to goal, kickstands down, and dump60. From there we need to hard code in move/turns to get it to the goal area on the floor. If isRedTeam is set to true at the beginning of the code, then once the sensor senses red on the ground it should put the kickstands up, same for blue